gcp/cloud_functions/function1/requirements.txt
# gcp/cloud_functions/function1/requirements.txt

Flask==2.2.5
psycopg2-binary==2.9.6
gunicorn==20.1.0
